backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic>=1.12.3
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
celery[redis]>=5.3.6
redis>=5.0.1
requests>=2.31.0
beautifulsoup4>=4.12.2
feedparser>=6.0.10
openai>=1.6.0
cryptography>=41.0.7
httpx>=0.25.2
slowapi>=0.1.9
lxml>=4.9.3
